feat : List<List<Boolean>> ladders 객체를 List<Line> lines로 변경
Ladder 객체의 기존 메서드 중 Line 객체가 처리하기에 더 적합하다고 판단한 메서드를 Line 객체 안으로 이동합니다.
